Fix engine controller autoloading
You can't route directly to application controller in an engine because
the application's application controller is loaded first and autoloading
doesn't trigger a second time for that constant for the engine; plus the
router isn't smart enough to know to try and autoloa a different one in
this isolated namespace.

So instead of trying to argue and fix it as a bug, just go with the flow
and don't route directly to application controller.