Add host type
Adds a host type with top-level host and hosts fields. This provides
the Taggable interface (hosts are taggable) and adds vms as a field of
hosts, so you can query the vms of a specified host or hosts.

AFAIK id, name and timestamps are the only fields guaranteed to be
non-null.