Add a DateTime type for coercing to/from ISO8601
Currently DateTimes are coerced to strings using the default
implementation of `Time#to_s`, which is the equivalent of calling
`time.strftime "%Y-%m-%d %H:%M:%S %z"`. While this is a fine
human-readable format, it would be better to render them in the
standard ISO8601 format for ease of parsing.

This change adds a new scalar type `DateTime`, which will coerce
DateTimes going out as ISO8601 encoded strings, and will expect
incoming times to be in the same format so that it can coerce them
back into Ruby times.