Implement Basic Authentication
This adds Basic Authentication around the current token authentication
mechanism.

The behavior can be roughly illustrated as follows:

1. If a user makes an unauthenticated request, they are challenged to
use basic authentication (this is conveyed via the status 401 and the
WWW-Authenticate header, which is required by RFC 2617)

2. If a user provides bad credentials, they are challenged in the same
way

3. If a user provides good credentials, they will be authenticated and
can proceed as normal

4. A user can still use token authentication as before, but will
not (currently) be challenged to use it in the case that they are not
successful. This behavior we will probably want to review.