Ensure RAILS_ENV is established before Rails is required
Since `spec_helper` loads Rails in order to create the GraphQL engine,
not having the `RAILS_ENV` value set at this point can lead to some
unfortunate behavior. It can, for instance, connect to the development
database (and this is never corrected), leading to test
failures. It'll also complain that there's no secrets file (as it
would with a brand new vanilla app). Setting it to "test" early on
mitigates both these issues.