Instrument all fields to camelcase
While it's not required by the GraphQL spec, it's an established
convention that GraphQL fields are defined in camelCase (and everything
is case sensitive).

Because we're in Rubyland, we're used to underscored things, and all of
the underlying methods that fields use are in underscores. To achieve
camelCase in the actual field name, you thus need to do one of the
following:

```ruby
field :someCoolData, :property => :some_cool_data
```

OR

```ruby
field :some_cool_property, :as => :someCoolData
```

To avoid having to add something like this for every single multiword
attribute in our models (and thus introducing an easy way to
accidentally introduce a field that we'd need to deprecate to correct),
FieldNameCamelizer is a instrumentation added to all fields which
converts all field names to camelCase if they aren't already.