Add `execute_graphql` integration macro
This consolidates on all the ceremony we have setting up a user token to
make a proper request in integration tests. It pairs nicely with
`as_user` but the only real requirement it has is to have a user defined
(either in a variable or memoized helper, whatever)

As there's plenty of implied magic happening, I've also added a custom
error to help newcomers find their way when trying to use it.