Move GraphQL Explorer JS to ui-classic
manageiq-graphql should *not* be a UI plugin.
It provides no screens, no toolbars or any such.

Until now, it did provide 1 react component for the UI to consume - the graphql explorer screen.
But the screen is development only, and multiple repositories requiring graphql is problematic - reducing complexity by making graphql not be a UI plugin.

https://bugzilla.redhat.com/show_bug.cgi?id=1704891