[rails6][gemspec] Update dependencies
- Updates `graphql-batch` to the latest version
- Use specific version of `graphql-preload`

The latter uses a fork of `graphql-preload` which will be hosted on
rubygems.manageiq.org, so we use a specific version and lock to that
since we control it.

It is a small light gem, so chances of it needing to change before a
substantial Rails upgrade are minimal and thus locking it to a specific
version shouldn't be a problem.